requirements.txt
Flask==3.1.1
flask-cors==6.0.1
geopy==2.4.1
numpy==2.3.2
pandas==2.3.1
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-dotenv==1.1.1
requests==2.32.4
